refactor(app): split HTTP types out of DebugRPCServer
Pure relocation: HTTPRequest/HTTPResponse move verbatim to their own
files. DebugRPCServer.swift sat at the 600-line file_length cap; the
split restores headroom the same way the AppAudioCapture+* line-cap
splits did. No behavior change.
