+++ app/MeetingTranscriber/Sources/DebugRPCServer.swift
@@ -483,89 +483,6 @@
         }
     }
 
-    // MARK: - HTTP types
-
-    struct HTTPRequest: Equatable {
-        let method: String
-        let path: String
-        let headers: [String: String]
-        let body: Data
-
-        init(method: String, path: String, headers: [String: String] = [:], body: Data = Data()) {
-            self.method = method
-            self.path = path
-            self.headers = headers
-            self.body = body
-        }
-
-        static func parse(_ data: Data) -> Self? {
-            let separator = Data("\r\n\r\n".utf8)
-            guard let separatorRange = data.range(of: separator) else { return nil }
-            let headerData = data.subdata(in: 0 ..< separatorRange.lowerBound)
-            guard let headerString = String(data: headerData, encoding: .utf8) else { return nil }
-            let lines = headerString.components(separatedBy: "\r\n")
-            guard let requestLine = lines.first else { return nil }
-            let parts = requestLine.split(separator: " ")
-            guard parts.count >= 2 else { return nil }
-            let method = String(parts[0])
-            let path = String(parts[1])
-
-            var headers: [String: String] = [:]
-            for line in lines.dropFirst() {
-                let pieces = line.split(separator: ":", maxSplits: 1, omittingEmptySubsequences: true)
-                guard pieces.count == 2 else { continue }
-                let key = pieces[0].trimmingCharacters(in: .whitespaces).lowercased()
-                let value = pieces[1].trimmingCharacters(in: .whitespaces)
-                headers[key] = value
-            }
-
-            let contentLength = headers["content-length"].flatMap(Int.init) ?? 0
-            let bodyStart = separatorRange.upperBound
-            let availableBody = data.count - bodyStart
-            guard availableBody >= contentLength else { return nil }
-            let body = data.subdata(in: bodyStart ..< bodyStart + contentLength)
-            return Self(method: method, path: path, headers: headers, body: body)
-        }
-    }
-
-    struct HTTPResponse {
-        let status: Int
-        let reason: String
-        let body: Data
-        let contentType: String
-
-        static func ok(body: Data, contentType: String) -> Self {
-            Self(status: 200, reason: "OK", body: body, contentType: contentType)
-        }
-
-        static func notFound() -> Self {
-            Self(status: 404, reason: "Not Found", body: Data(), contentType: "text/plain")
-        }
-
-        static func unauthorized() -> Self {
-            Self(status: 401, reason: "Unauthorized", body: Data(), contentType: "text/plain")
-        }
-
-        static func forbidden() -> Self {
-            Self(status: 403, reason: "Forbidden", body: Data(), contentType: "text/plain")
-        }
-
-        static func badRequest() -> Self {
-            Self(status: 400, reason: "Bad Request", body: Data(), contentType: "text/plain")
-        }
-
-        func serialize() -> Data {
-            var response = "HTTP/1.1 \(status) \(reason)\r\n"
-            response += "Content-Type: \(contentType)\r\n"
-            response += "Content-Length: \(body.count)\r\n"
-            response += "Connection: close\r\n"
-            response += "\r\n"
-            var out = Data(response.utf8)
-            out.append(body)
-            return out
-        }
-    }
-
     // MARK: - Speaker DB action types
 
     enum SpeakerActionOutcome: String {
